cleaned up code

+++ commands/mod/prune.js
@@ -10,11 +10,10 @@ module.exports = class PurgeCommand extends Command {
       memberName: 'prune',
       description: 'Deletes messages.',
       examples: ['prune 4'],
-      clientPermissions: ['MANAGE_MESSAGES'],
       args: [{
           key: 'number',
           prompt: 'How many messages to purge?',
-          type: 'integer',
+          type: 'integer'
       }]
     });
   }
@@ -26,13 +25,16 @@ module.exports = class PurgeCommand extends Command {
       return message.channel.send('Please prune between 2 and 100 messages.')
     }
 
+    if (!message.channel.permissions.has(this.client.user).has('MANAGE_MESSAGES')) {
+      return message.channel.send("Help! I don't have permission to manage messages!")
+    }
+
     if (!message.member.permissions.has('MANAGE_MESSAGES')) {
       return message.channel.send("You don't have permission to manage messages.")
     }
 
     message.channel.fetchMessages({limit: number}).then(messages => { 
       message.channel.bulkDelete(messages)
-      message.channel.send(`Removed ${messages.size} messages.`)
-   })
+      message.channel.send(`Removed ${messages.size} messages.`)})
   }; //end run
 }; //end command
